feat: moved helper functions to separate files
