+++ libs/core/src/lib/utils/decorators/apply-css-class.decorator.ts
@@ -1,20 +1,4 @@
-import { ELEMENT_REF_EXCEPTION } from '../public_api';
-
-function uuidv4(): string {
-    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-        // tslint:disable-next-line: no-bitwise
-        const r = (Math.random() * 16) | 0,
-            // tslint:disable-next-line: no-bitwise
-            v = c === 'x' ? r : (r & 0x3) | 0x8;
-        return v.toString(16);
-    });
-}
-
-function buildClassFromMap(nativeElement: { _classMap: any }): string {
-    return Object.keys(nativeElement._classMap)
-        .map((key) => nativeElement._classMap[key])
-        .reduce((acc, value) => (acc += value), '');
-}
+import { ELEMENT_REF_EXCEPTION, getStringFromHashMap, uuidv4 } from '../public_api';
 
 /**
  * Method decorator to apply css class to a component through native element
@@ -46,7 +30,9 @@ export function applyCssClass(target: any, propertyKey: string, descriptor: Prop
 
             elementRef.nativeElement._classMap[this._uuidv4] = `${_class} ${this.class ? this.class : ''}`;
 
-            (elementRef.nativeElement as HTMLElement).className = buildClassFromMap(elementRef.nativeElement);
+            (elementRef.nativeElement as HTMLElement).className = getStringFromHashMap<string>(
+                elementRef.nativeElement._classMap
+            );
         }
 
         return _class;
