+++ libs/core/src/lib/utils/functions/uuidv4-generator.ts
@@ -1,0 +1,9 @@
+export function uuidv4(): string {
+    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
+        // tslint:disable-next-line: no-bitwise
+        const r = (Math.random() * 16) | 0,
+            // tslint:disable-next-line: no-bitwise
+            v = c === 'x' ? r : (r & 0x3) | 0x8;
+        return v.toString(16);
+    });
+}
